livraria/urls.py
from django.contrib import admin
from django.urls import include, path

from core.views import CategoriaViewSet, EditoraViewSet, AutorViewSet, LivroViewSet

from rest_framework.routers import DefaultRouter

from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
)

router = DefaultRouter()
router.register(r'categorias', CategoriaViewSet)
router.register(r'editoras', EditoraViewSet)
router.register(r'autores', AutorViewSet)
router.register(r'livros', LivroViewSet)

urlpatterns = [
    #Admin
    path("admin/", admin.site.urls),
    #SimpleJWT
    path('token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    #API
    path('', include(router.urls)),
]

